+++ PyMOTW/subprocess/subprocess_signal_setsid.py
@@ -1,0 +1,37 @@
+#!/usr/bin/env python
+# encoding: utf-8
+#
+# Copyright (c) 2009 Doug Hellmann All rights reserved.
+#
+"""
+"""
+#end_pymotw_header
+
+import os
+import signal
+import subprocess
+import tempfile
+import time
+import sys
+
+script = '''#!/bin/sh
+echo "Shell script in process $$"
+set -x
+python signal_child.py
+'''
+script_file = tempfile.NamedTemporaryFile('wt')
+script_file.write(script)
+script_file.flush()
+
+proc = subprocess.Popen(['sh %s' % script_file.name], 
+                        shell=True, 
+                        close_fds=True,
+                        preexec_fn=os.setsid,
+                        )
+print 'PARENT: Pausing before sending signal to child %s...' % proc.pid
+sys.stdout.flush()
+time.sleep(1)
+print 'PARENT: Signaling process group %s' % proc.pid
+sys.stdout.flush()
+os.killpg(proc.pid, signal.SIGUSR1)
+time.sleep(3)